+++ Installer/install.py
@@ -1,44 +1,52 @@
 import subprocess, platform
 
+def installProgram(cmd):
+    p = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+    p.wait()
+    return p.returncode
+
+def getRequirements():
+    p = subprocess.Popen("pip3 install -r -y requirements.txt", stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+    p.wait()
+    return p.returncode
+
+def searchFor(program_name):
+    p = subprocess.Popen(program_name, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+    p.wait()
+    return p.returncode
+
 def collectPackages():
     print("Collecting packages...")
     if(platform.system() == "Linux"):
-        p = subprocess.Popen("sudo pip3 install -r requirements.txt", stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-        p.wait()
-        if(p.returncode != 1):
+        p = getRequirements()
+        
+        if(p != 0):
             print("Pip is not installed.\nInstalling...")
-            p = subprocess.Popen("sudo apt-get install python3-pip -y", stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-            p.wait()
+            installProgram("sudo apt-get install python3-pip -y")
             print("Pip installed\n")
-            p = subprocess.Popen("sudo pip3 install -r requirements.txt", stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-            p.wait()
-            collectPackages()
+            getRequirements()
+
         print("Searching for git...")
-        p = subprocess.Popen("git", stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-        p.wait()
-        if(p.returncode != 1):
+        p = searchFor("git")
+
+        if(p != 0):
             print("Installing git...")
-            p = subprocess.Popen("sudo apt-get install git -y", stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-            p.wait()
+            installProgram("sudo apt-get install git -y")
             print("Git installed\n")
         else:
             print("Git found")
-        from rsc.install import fetchData, install
-        fetchData()
-        install()
+
     if(platform.system() == "Windows"):
-        p = subprocess.Popen("pip3 install -r -y requirements.txt", stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-        p.wait()
+        getRequirements()
         
-        p = subprocess.Popen("git", stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-        p.wait()
-        if(p.returncode == -1):
-            print("Es wurde kein git gefunden... entweder installierst du es oder du wartest noch bis der Installer auch die Gitinstallation auf Windows unterstützt...")
+        p=searchFor("git")
+        if(p != 1):
+            print("Es wurde kein git gefunden... entweder installierst du es oder du wartest noch bis der Installer auch die GIT-Installation auf Windows unterstützt...")
             exit()
     print("Collected all packages.")
 
 if __name__ == "__main__":
     collectPackages()
-    from rsc.install import fetchData, install
+    from rsc.source import fetchData, install
     fetchData()
     install()
